cleaned code about transfers and removed unused Transactor classd
