+++ Transaction.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Globalization;
 
 namespace Project_2
 {
@@ -9,111 +10,107 @@ namespace Project_2
         protected double Amount { get; set; }
         protected Account DestinationAccount { get; set; }
 
-        public abstract void DoTransferFromCurrentAccountToSavingAccountAccordingToDate(int debitClient_id, int SavingAccount_id, double amount, DateTime transferDate);
-        public abstract void DoTransferFromCurrentToOtherCurrentAccountAccordingToDate(int debitClient_id, int clientIdOfExternalAccount, double amount, DateTime transferDate);
-        public abstract void DoTransferFromSavingToCurrentAccountAccordingToDate(int debitClient_id, int debitSavingAccount_id, int recipientAccount_id, double amount, DateTime transferDate);
-        
 
 
-        public char ChooseDebitAccount()
+        public void QueryTransferFromCurrentToCurrent(int creditCurrentAccount_id, double amount, DateTime executionDate)
         {
-            Console.WriteLine("Specify from which account you want to transfer money:");
-            Console.WriteLine("Current account (c) or saving account (s)");
-            char debitAccount = Convert.ToChar(Console.ReadLine());
-
-            return debitAccount;
-        }
-
-        public char ChooseRecipientAccount(int debitClient_id)
-        {
-            Console.WriteLine("Specify to which account you want to transfer money:");
-            Console.WriteLine("Your current account (c), one of your saving account (s), or to a foreign current account (e).");
-            char creditAccount = Convert.ToChar(Console.ReadLine());
-
-                return creditAccount;
-        }
-
-        public int GetAccountIdFromAccountType(int debitClient_id, char recipientAccount)
-        {
-            if (recipientAccount == 'c')
-            {
-                int currentAccount_id = 0;
-                return currentAccount_id;
-            }
-            else if (recipientAccount == 's')
-            {
-                int savingAccount_id = Transactor.GetSavingAccountIdFromClientChoice(debitClient_id);
-                return savingAccount_id;
-            }
-            else if (recipientAccount == 'e')
-            {
-                Console.WriteLine("Specify the id number of the beneficiary account");
-                int externalAccount_id = Convert.ToInt32(Console.ReadLine());
-                
-                int clientIdOfExternalAccount = GetClientIDFromAccountId(externalAccount_id);
-                if (clientIdOfExternalAccount > 0)
-                {
-                    return clientIdOfExternalAccount;
-                }
-                else
-                {
-                    throw new ArgumentException("Unvalid account number");
-                }
-            }
-            else
-            {
-                throw new ArgumentException("Exiting program due to input error");
-            }
-        }
-
-        
-        public void ExecuteTransfer(double amount, DateTime transferDate, char debitAccount, int debitSavingAccount_id, char recipientAccount, int recipientAccount_id)
-        {
-            
-            // Récupère l'ID du client dans la propriété qui doit être définie lors de la connexion du client sur l'interface
             int debitClient_id = Person.ID;
+            string checkCurrentAccountContent = $"SELECT amount FROM CurrentAccounts WHERE client_id = {debitClient_id}";
+            decimal CurrentAccountContent = ConnectionDB.ReturnDecimal(checkCurrentAccountContent);
+            string getCurrentAccountOverdraft = $"SELECT overdraft FROM CurrentAccounts WHERE client_id = {debitClient_id}";
+            decimal CurrentAccountOverdraft = ConnectionDB.ReturnDecimal(getCurrentAccountOverdraft);
 
-            if (debitAccount == 'c')
+            if (Convert.ToDouble(CurrentAccountContent + CurrentAccountOverdraft) > amount)
             {
-                TransferFromCurrentToUserChoice(debitClient_id, amount, transferDate, recipientAccount, recipientAccount_id);
-            }
-            else if (debitAccount == 's')
-            {
-                DoTransferFromSavingToCurrentAccountAccordingToDate(debitClient_id, debitSavingAccount_id, recipientAccount_id, amount, transferDate);
-            }
-            else
-            {
-                Console.WriteLine("Exiting program due to input error");
+                string queryString =
+                                 $"INSERT INTO \"Transaction\" (currentAccount_id, transactionType, beneficiaryCurrentAccount_id, amount, executionDate, status) " +
+                                 $"VALUES (" +
+                                 $"(SELECT id FROM CurrentAccounts WHERE client_id = {debitClient_id}), " +
+                                 $"\'Money Transfer\', " +
+                                 $"{creditCurrentAccount_id}, " +
+                                 $"{amount}, " +
+                                 $"\'{executionDate}\'," +
+                                 $"\'pending\');";
+                ConnectionDB.NonQuerySQL(queryString);
             }
         }
 
-
-        public void TransferFromCurrentToUserChoice(int debitClient_id, double amount, DateTime transferDate, char recipientAccount, int recipientAccount_id)
+        public void QueryTransferFromSavingToCurrent(int debitSavingAccount_id, double amount, DateTime executionDate)
         {
-            if (recipientAccount == 's')
+            int debitClient_id = Person.ID;
+            string checkSavingAccountContent = $"SELECT amount FROM SavingAccounts WHERE id = {debitSavingAccount_id}";
+            decimal SavingAccountContent = ConnectionDB.ReturnDecimal(checkSavingAccountContent);
+
+            if ((Convert.ToDouble(SavingAccountContent) - amount) >= 0)
             {
-                //int SavingAccount_id = Transactor.GetSavingAccountIdFromClientChoice(debitClient_id);
-                DoTransferFromCurrentAccountToSavingAccountAccordingToDate(debitClient_id, recipientAccount_id, amount, transferDate);
+                string queryString =
+                                 $"INSERT INTO \"Transaction\" (savingAccount_id, transactionType, beneficiaryCurrentAccount_id, amount, executionDate, status) " +
+                                 $"VALUES (" +
+                                 $"{debitSavingAccount_id}, " +
+                                 $"\'Money Transfer\', " +
+                                 $"(SELECT id FROM CurrentAccounts WHERE client_id = {debitClient_id}), " +
+                                 $"{amount}, " +
+                                 $"\'{executionDate}\'," +
+                                 $"\'pending\');";
+                ConnectionDB.NonQuerySQL(queryString);
             }
-            else if (recipientAccount == 'e')
+        }
+
+        public void QueryTransferFromCurrentToSaving(int SavingAccount_id, double amount, DateTime firstExecution)
+        {
+            int debitClient_id = Person.ID;
+            string checkCurrentAccountContent = $"SELECT amount FROM CurrentAccounts WHERE client_id = {debitClient_id}";
+            decimal CurrentAccountContent = ConnectionDB.ReturnDecimal(checkCurrentAccountContent);
+            string getCurrentAccountOverdraft = $"SELECT overdraft FROM CurrentAccounts WHERE client_id = {debitClient_id}";
+            decimal CurrentAccountOverdraft = ConnectionDB.ReturnDecimal(getCurrentAccountOverdraft);
+
+            string checkSavingAccountContent = $"SELECT amount FROM SavingAccounts WHERE id = {SavingAccount_id}";
+            decimal SavingAccountContent = ConnectionDB.ReturnDecimal(checkSavingAccountContent);
+            string checkSavingAccountCeiling = $"SELECT ceiling FROM SavingAccounts WHERE id = {SavingAccount_id}";
+            decimal SavingAccountCeiling = ConnectionDB.ReturnDecimal(checkSavingAccountCeiling);
+
+            if (Convert.ToDouble(CurrentAccountContent - CurrentAccountOverdraft) >= amount)
             {
-                /*Console.WriteLine("Specify the id number of the beneficiary account");
-                int externalAccount_id = Convert.ToInt32(Console.ReadLine());*/
-                
-                if (recipientAccount_id > 0)
+                if (((Convert.ToDouble(SavingAccountContent)) + amount) > Convert.ToDouble(SavingAccountCeiling))
                 {
-                    DoTransferFromCurrentToOtherCurrentAccountAccordingToDate(debitClient_id, recipientAccount_id, amount, transferDate);
+                    string queryString =
+                                 $"INSERT INTO \"Transaction\" (currentAccount_id, transactionType, beneficiaryAccount_id, amount, executionDate, status) " +
+                                 $"VALUES (" +
+                                 $"(SELECT id FROM CurrentAccounts WHERE client_id = {debitClient_id}), " +
+                                 $"\'Money Transfer\', " +
+                                 $"(SELECT id FROM SavingAccounts WHERE client_id = {debitClient_id}), " +
+                                 $"{amount}, " +
+                                 $"\'{firstExecution}\'," +
+                                 $"\'pending\');";
+                    ConnectionDB.NonQuerySQL(queryString);
                 }
                 else
                 {
-                    Console.WriteLine("Please specify a valid recipient account number.");
+                    Console.WriteLine($"Impossible transfer. Transfer would exceed ceiling limit.");
                 }
             }
             else
             {
-                Console.WriteLine("Exiting program due to input error");
+                Console.WriteLine($"There is not enough money on current account to perform transfer");
             }
         }
+
+        public static DateTime CheckDate(string userInput)
+        {
+            CultureInfo fr = new CultureInfo("fr");
+            DateTime dateValue;
+            if (!DateTime.TryParseExact(userInput, "d", fr, DateTimeStyles.AllowLeadingWhite, out dateValue))
+            {
+                throw new ArgumentException("Unvalid date format. Accepted format dd/mm/yyyy.");
+            }
+            else
+            {
+                dateValue = Convert.ToDateTime(userInput);
+            }
+
+            return dateValue;
+        }
+
 
         public int GetClientIDFromAccountId(int account_id)
         {
